Update README.md
Team name: ECoders
Members:
1.Abel Sabu Payyanickal, Email: asp2003@my.yorku.ca (Section M Lab 03),
2.Jeremy John Varughese, Email: jeremyj3@my.yorku.ca (Section M Lab 04),
3.Uzair Ansari, Email: uzairaa@my.yorku.ca (Section M Lab 04),
4.Vania Sameen; Email: vania7@my.yorku.ca (Section M Lab 01),

Project Name: Rock, Paper, Scissors.

Project Description:
We have settled on Rock, Paper, Scissors, a game in which the player selects a symbol by clicking on an image of it, and then it shows whether the user or the server won. Best out of 5 (single player) and 1v1 (multiplayer) will be completed for this project. Players will also have the option to play with thematic variations of Rock,Paper,Scissors such as Cat,Mouse,Dog or Earth,Fire,Water . This will be available on the settings page . To achieve the desired result, we intend to use Visual Studio Code to execute Java script to control the behavior of the webpage (Active Content), Html Text Editor to give the webpage structure, and CSS to design the webpage.

Functional Requirements:

• In the homepage we will have a yellow background for the whole page where It’s written Rock Paper Scissors and there is a begin button on a blue box.

• When the user clicks begin the next page comes which says to Select Mode (Single Player, Multiplayer) and Select Rounds (Best of 3, Best of 5, Best of 7).

• When the user clicks single player you enter the single player mode.

• When the user clicks multiplayer you enter the multiplayer mode.

• The user should click the number of rounds before playing or by default the round is set to 
  
•The user should click the settings button and be able to choose one of the three thematic variations of the game 

• If the user selects Single Player, then the user will be asked to select and click a figure out of three figures which include a rock, a paper and a scissor.

• The other part of the screen which is played by the computer will randomly generate an output and it will be shown.

• Followed by the RESULT page where the score and an option to play again will be shown.

• If the user selects Multiplayer, then the user/users (Player 1 & Player 2) will be asked to select and click a figure out of three figures which include a rock, a paper and a scissor.

• Followed by the RESULTS page where the score and an op